add cc31
