+++ README.md
@@ -1,20 +1,11 @@
 # data
-## Hash Tables.
-Implement a Hashtable with the following methods:
-add: takes in both the key and value. This method should hash the key, and add the key and value pair to the table, handling collisions as needed.
-get: takes in the key and returns the value from the table.
-contains: takes in the key and returns a boolean, indicating if the key exists in the table already.
-hash: takes in an arbitrary key and returns an index in the collection.
+## Repeated-word.
+Write a function that accepts a lengthy string parameter to Find the first repeated word in a book.
+Without utilizing any of the built-in library methods available to your language, return the first word to occur more than once in that provided string.
 ## Approach & Efficiency
-Implement a Hashtable with the following methods:
-add: takes in both the key and value. This method should hash the key, and add the key and value pair to the table, handling collisions as needed.
-get: takes in the key and returns the value from the table.
-contains: takes in the key and returns a boolean, indicating if the key exists in the table already.
-hash: takes in an arbitrary key and returns an index in the collection.
+Write a function that accepts a lengthy string parameter to Find the first repeated word in a book.
+Without utilizing any of the built-in library methods available to your language, return the first word to occur more than once in that provided string.
 ## Big O
-- add :Time o(1), space o(1).
-- get :Time o(n), space o(1).
-- contains :Time o(n), space o(1).
-- hash " Time o(n), space o(1).
+add :Time o(n) there is for loop, space o(n)because i declare an array by split the string by the space.
 ## Solution
-![Solution](/assets/cc30.jpeg)
+![Solution](/assets/cc31.jpeg)
